Redaction: Move database to separate module
Having a separate module for the database operations will let use reason
about the Redaction script as a standalone application that *only* uses
the library code. This is because this script should exclusively focus
on interfacing the library with the user code.
